DB関連ファイルを分離 & マイグレーションをGRDBのやつを使うように
