+++ iMast/SharedSources/MastodonModel/MastodonUserToken.swift
@@ -1,9 +1,10 @@
 //
-//  MastodonModel.swift
-//  iMast
+//  MastodonUserToken.swift
 //
-//  Created by rinsuki on 2017/04/24.
-//  
+//  iMast https://github.com/cinderella-project/iMast
+//
+//  Created by user on 2019/08/24.
+//
 //  ------------------------------------------------------------------------
 //
 //  Copyright 2017-2019 rinsuki and other contributors.
@@ -19,236 +20,12 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-//
 
 import Foundation
-import Alamofire
 import SwiftyJSON
-import GRDB
+import Alamofire
 import Hydra
-
-let fileURL = getFileURL()
-let dbQueue = try! DatabaseQueue(path: fileURL.path)
-func getFileURL() -> URL {
-    let oldFileURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.jp.pronama.imast")!.appendingPathComponent("imast.sqlite")
-    let fileURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.jp.pronama.imast")!.appendingPathComponent("Library/imast.sqlite")
-    if !FileManager.default.fileExists(atPath: fileURL.path) && FileManager.default.fileExists(atPath: oldFileURL.path) {
-        print("migrate: AppGroup -> AppGroup Library")
-        try! FileManager.default.moveItem(atPath: oldFileURL.path, toPath: fileURL.path)
-    }
-    if !FileManager.default.fileExists(atPath: fileURL.path) && FileManager.default.fileExists(atPath: NSHomeDirectory()+"/Library/imast.sqlite") {
-        print("migrate: App Library -> AppGroup")
-        try! FileManager.default.moveItem(atPath: NSHomeDirectory()+"/Library/imast.sqlite", toPath: fileURL.path)
-    }
-    return fileURL
-}
-func initDatabase() {
-    try! dbQueue.inDatabase { db in
-        print("--start db migrate")
-        try db.execute("CREATE TABLE IF NOT EXISTS app (id text primary key, client_id text, client_secret text, redirect_uri text, instance_hostname text, name text)", arguments: nil)
-        try db.execute("CREATE TABLE IF NOT EXISTS user (id text primary key, access_token text, instance_hostname text, app_id text, name text, screen_name text, avatar_url text, last_used int)", arguments: nil)
-        print("--end db migrate")
-    }
-}
-
-var mastodonInstanceInfoCache: [String: JSON] = [:]
-
-public class MastodonInstance {
-    var hostName: String
-    var name: String?
-    var description: String?
-    var email: String?
-    var url: URL {
-        return URL(string: "https://\(self.hostName)")!
-    }
-        
-    init(hostName: String = "mastodon.social") {
-        self.hostName = hostName.pregReplace(pattern: ".+\\@", with: "").lowercased()
-    }
-    
-    func getInfo() -> Promise<JSON> {
-        return Promise<JSON> { resolve, reject, _ in
-            if let cache = mastodonInstanceInfoCache[self.hostName] {
-                resolve(cache)
-                return
-            }
-            Alamofire.request("https://\(self.hostName)/api/v1/instance").responseJSON { res in
-                // print(res)
-                if let error = res.error {
-                    reject(error)
-                    return
-                }
-                if res.result.value == nil {
-                    reject(APIError.nil("response.result.value"))
-                    return
-                }
-                let json = JSON(res.result.value!)
-                self.name = json["name"].string
-                self.description = json["description"].string
-                self.email = json["email"].string
-                mastodonInstanceInfoCache[self.hostName] = json
-                resolve(json)
-            }
-        }
-    }
-    
-    func createApp(name: String = "iMast", redirect_uri: String = "imast://callback/") -> Promise<MastodonApp> {
-        return Promise<MastodonApp> { resolve, reject, _ in
-            let params = [
-                "client_name": name,
-                "scopes": "read write follow",
-                "redirect_uris": redirect_uri,
-                "website": "https://cinderella-project.github.io/iMast/",
-            ]
-            Alamofire.request("https://\(self.hostName)/api/v1/apps", method: .post, parameters: params).responseJSON { res in
-                if let error = res.error {
-                    reject(error)
-                    return
-                }
-                if res.result.value == nil {
-                    reject(APIError.nil("response.result.value"))
-                    return
-                }
-                let json = JSON(res.result.value!)
-                print(json)
-                resolve(MastodonApp(instance: self, info: json, name: name, redirectUri: redirect_uri))
-            }
-        }
-    }
-}
-
-public class MastodonApp {
-    var clientId: String
-    var clientSecret: String
-    var name: String
-    var redirectUri: String
-    var instance: MastodonInstance
-    var id: String
-    
-    init(instance: MastodonInstance, info: JSON, name: String, redirectUri: String) {
-        self.instance = instance
-        print(info)
-        clientId = info["client_id"].stringValue
-        clientSecret = info["client_secret"].stringValue
-        self.name = name
-        self.redirectUri = redirectUri
-        self.id = genRandomString()
-    }
-    
-    init(instance: MastodonInstance, clientId: String, clientSecret: String, name: String, redirectUri: String) {
-        self.instance = instance
-        self.clientId = clientId
-        self.clientSecret = clientSecret
-        self.name = name
-        self.redirectUri = redirectUri
-        self.id = genRandomString()
-    }
-    static func initFromId(appId: String) -> MastodonApp {
-        return try! dbQueue.inDatabase { db in
-            let row = try Row.fetchOne(db, "SELECT * from app where id=? LIMIT 1", arguments: [appId])!
-            return initFromRow(row: row)
-        }
-    }
-    static func initFromRow(row: Row) -> MastodonApp {
-        let app = MastodonApp(
-            instance: MastodonInstance(hostName: row["instance_hostname"]),
-            clientId: row["client_id"],
-            clientSecret: row["client_secret"],
-            name: row["name"],
-            redirectUri: row["redirect_uri"]
-        )
-        app.id = row["id"]
-        return app
-    }
-    
-    @discardableResult
-    func save() -> Bool {
-        do {
-            try dbQueue.inDatabase { db in
-                try db.execute("INSERT OR REPLACE INTO app (id, client_id, client_secret, name, redirect_uri, instance_hostname) VALUES (?,?,?,?,?,?)", arguments: [
-                    self.id,
-                    self.clientId,
-                    self.clientSecret,
-                    self.name,
-                    self.redirectUri,
-                    self.instance.hostName,
-                ])
-            }
-            return true
-        } catch {
-            print(error)
-            return false
-        }
-    }
-    
-    func getAuthorizeUrl() -> String {
-        return "https://\(instance.hostName)/oauth/authorize?client_id=\(clientId)&redirect_uri=\(self.redirectUri)&scope=read+write+follow&response_type=code&state="+self.id
-    }
-    
-    func authorizeWithCode(code: String) -> Promise<MastodonUserToken> {
-        return Promise<MastodonUserToken> { resolve, reject, _ in
-            Alamofire.request("https://\(self.instance.hostName)/oauth/token", method: .post, parameters: [
-                "grant_type": "authorization_code",
-                "redirect_uri": self.redirectUri,
-                "client_id": self.clientId,
-                "client_secret": self.clientSecret,
-                "code": code,
-                "state": self.id,
-            ]).responseJSON { res in
-                if res.result.value == nil {
-                    reject(APIError.nil("res.response.value"))
-                    return
-                }
-                if res.result.value == nil {
-                    reject(APIError.nil("response.result.value"))
-                    return
-                }
-                let json = JSON(res.result.value!)
-                print(json)
-                resolve(MastodonUserToken(app: self, token: json["access_token"].stringValue))
-            }
-        }
-    }
-    
-    func authorizeWithPassword(email: String, password: String) -> Promise<MastodonUserToken> {
-        return Promise { resolve, reject, _ in
-            Alamofire.request("https://\(self.instance.hostName)/oauth/token", method: .post, parameters: [
-                "grant_type": "password",
-                "username": email,
-                "password": password,
-                "client_id": self.clientId,
-                "client_secret": self.clientSecret,
-                "scope": "read write follow",
-            ]).responseJSON { response in
-                if ((response.response?.url?.absoluteString) ?? "").contains("/auth/sign_in") { // MastodonのAPIはクソ!w
-                    reject(APIError.errorReturned(errorMessage: "Emailかパスワードが誤っています", errorHttpCode: -1))
-                    return
-                }
-                if response.result.value == nil {
-                    reject(APIError.nil("response.result.value"))
-                    return
-                }
-                var json = JSON(response.result.value!)
-                json["_response_code"].int = response.response?.statusCode ?? 599
-                print(json)
-                if json["_response_code"].intValue >= 400 {
-                    if let error = json["error"].string {
-                        reject(APIError.errorReturned(errorMessage: error, errorHttpCode: json["_response_code"].intValue))
-                        return
-                    } else {
-                        reject(APIError.unknownResponse(errorHttpCode: json["_response_code"].intValue))
-                        return
-                    }
-                }
-                if json["access_token"].string == nil {
-                    reject(APIError.nil("access_token"))
-                    return
-                }
-                resolve(MastodonUserToken(app: self, token: json["access_token"].stringValue))
-            }
-        }
-    }
-}
+import GRDB
 
 public class MastodonUserToken: Equatable {
     var id: String?
@@ -565,96 +342,4 @@ public class MastodonUserToken: Equatable {
     public static func == (lhs: MastodonUserToken, rhs: MastodonUserToken) -> Bool {
         return lhs.id == rhs.id
     }
-}
-
-var imageCache: [String: UIImage] = [:]
-var imageResizeCache: [String: [Int: UIImage]] = [:]
-
-@available(*, unavailable)
-public func getImage(url: String, size: Int = -1) -> Promise<UIImage> {
-    return Promise(in: .background) { resolve, reject, _ in
-        let resizedImagePath = NSHomeDirectory() + "/Library/Caches/image/" + url.sha256 + "_resize_\(String(size))_scale_"+UIScreen.main.scale.description
-        if size>0 {
-            if let resizedImageCache = imageResizeCache[url]?[size] {
-                resolve(resizedImageCache)
-                return
-            }
-            if FileManager.default.fileExists(atPath: resizedImagePath) {
-                print("Get From Storage(with resized image):"+url)
-                do {
-                    if let resizedCacheImage = UIImage(data: try Data(contentsOf: URL(fileURLWithPath: resizedImagePath), options: NSData.ReadingOptions.mappedIfSafe)) {
-                        if imageResizeCache[url] == nil {
-                            imageResizeCache[url] = [:]
-                        }
-                        imageResizeCache[url]![size] = resizedCacheImage
-                        resolve(resizedCacheImage)
-                        return
-                    }
-                } catch {
-                    print(error)
-                    
-                }
-            }
-        }
-        if let cacheImage = imageCache[url] {
-            resolve(cacheImage)
-            return
-        }
-        var img: UIImage!
-        var imgData = Data()
-        var isInternet = false
-        if FileManager.default.fileExists(atPath: NSHomeDirectory() + "/Library/Caches/image/" + url.sha256) {
-            print("Get From Storage:"+url)
-            img = UIImage(data: (try? Data(contentsOf: URL(fileURLWithPath: NSHomeDirectory() + "/Library/Caches/image/" + url.sha256), options: NSData.ReadingOptions.mappedIfSafe)) ?? Data())
-            if img == nil {
-                try? FileManager.default.removeItem(atPath: NSHomeDirectory() + "/Library/Caches/image/" + url.sha256)
-            }
-        }
-        if img == nil {
-            print("Get From Server:"+url)
-            imgData = try Data(contentsOf: URL(string: url)!, options: NSData.ReadingOptions.mappedIfSafe)
-            img = UIImage(data: imgData)
-            isInternet = true
-            do {
-                if !FileManager.default.fileExists(atPath: NSHomeDirectory() + "/Library/Caches/image") {
-                    try FileManager.default.createDirectory(atPath: NSHomeDirectory() + "/Library/Caches/image", withIntermediateDirectories: false, attributes: nil)
-                }
-            } catch {
-                print(error)
-            }
-        }
-        if img == nil {
-            print("Error Image Decode Failed:"+url)
-            reject(APIError.decodeFailed)
-            return
-        }
-        if isInternet {
-            try? imgData.write(to: URL(fileURLWithPath: NSHomeDirectory() + "/Library/Caches/image/" + url.sha256))
-        }
-        imageCache.updateValue(img!, forKey: url)
-        if size > 0 {
-            if let resizedImageCache = imageResizeCache[url]?[size] {
-                img = resizedImageCache
-            } else {
-                let resizedSize = CGSize(width: size, height: size)
-                UIGraphicsBeginImageContextWithOptions(resizedSize, false, UIScreen.main.scale)
-                img?.draw(in: CGRect(origin: .zero, size: resizedSize))
-                img = UIGraphicsGetImageFromCurrentImageContext()
-                UIGraphicsEndImageContext()
-                if imageResizeCache[url] == nil {
-                    imageResizeCache[url] = [:]
-                }
-                imageResizeCache[url]![size] = img
-            }
-        }
-        resolve(img!)
-    }
-}
-    
-public func genRandomString() -> String {
-    return (Date().timeIntervalSince1970*1000).description.sha256 + (arc4random().description.sha256)
-}
-
-public func genState(hostName: String) {
-    return
 }
